+++ src/components/content/museum-display/actionsTypes.js
