+++ src/components/content/museum-display/reducers.js
