+++ src/store/reducer.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 
-import museumsReducer from '../components/content/reducers'
+import museumsReducer from '../components/content/museum-display/reducers'
 
 
 const rootReducer = combineReducers({
